# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Japan Eco Tracking API'
  version: 1.0.0
paths:
...
  '/api/v1/members/{member_id}/stamps':
    get:
      tags:
        - Members
      summary: 'Get stamp list of user'
      description: 'Get stamp list of user'
      operationId: 'Get all'
      parameters:
        -
          name: member_id
          in: path
          description: 'ID of member to return'
          required: true
          schema:
            type: integer
            format: int64
            example: 1
        -
          name: Authorization
          in: header
          required: true
          schema:
            type: string
          example: 76f58b2747fa65b4dc157f9c1c9bbff708bffac652d8f67a60cff95ef4faf956e2767846b811af4133e277a1664f195853864b911c87c3c6fcf4e6bcccaedb86
      responses:
        200:
          description: 'Get stamp list success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Get stamp list success'
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: string
                          example: 1
                        name:
                          type: string
                          example: eos
                        description:
                          type: text
                          example: ut
                        latitude:
                          type: float
                          example: 31.416663
                        longitude:
                          type: number
                          example: 130.113251
                        thumbnail:
                          type: string
                          example: 'http://localhost/uploads/400x300.jpg'
                        type:
                          type: string
                          example: sint
                        created_at:
                          type: string
                          example: '2020-12-02T17:21:36.000000Z'
                        member:
                          properties:
                            id:
                              type: string
                              example: 1
                            montbell_login_user_id:
                              type: string
                              example: muratatest5
                            created_at:
                              type: string
                              example: '2020-12-02T17:21:36.000000Z'
                            updated_at:
                              type: string
                              example: '2020-12-04T06:26:04.000000Z'
                          type: object
                        route:
                          properties:
                            id:
                              type: string
                              example: 1
                            area_id:
                              type: number
                              example: 1
                            number:
                              type: number
                              example: 1
                            name:
                              type: string
                              example: example
                            description:
                              type: text
                              example: description
                            movement:
                              type: string
                              example: drive
                            stamina_level:
                              type: number
                              example: 1
                            range:
                              type: float
                              example: 8.5
                            max_elevation:
                              type: float
                              example: 1000.5
                            journey_time:
                              type: number
                              example: 60.5
                            line_color:
                              type: string
                              example: red
                            geometry:
                              type: array
                              items:
                                type: array
                                items:
                                  type: number
                                  example: ''
                                example:
                                  - 130.111679
                                  - 31.421721
                            point_center:
                              type: array
                              items:
                                type: number
                                example: ''
                              example:
                                - 130.111679
                                - 31.421721
                            zoom_level:
                              type: float
                              example: 8.5
                          type: object
                      type: object
                type: object
...
  /api/v1/news:
    get:
      tags:
        - News
      summary: 'Get all news'
      description: 'Get all news'
      operationId: 'Get all'
      parameters:
        -
          name: order_by
          in: query
          description: 'Sort order. Default to release_time descending if not set.'
          schema:
            type: string
            example: 'release_time,desc'
        -
          name: limit
          in: query
          description: 'Max items to return. Default to paging if not set.'
          schema:
            type: string
            format: int
            example: 5
        -
          name: page
          in: query
          description: 'Page number to return'
          schema:
            type: string
            format: int
            example: 1
      responses:
        200:
          description: 'Get all news success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Get all news success'
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: number
                          example: 1
                        member_id:
                          type: number
                          example: 1
                        news_id:
                          type: text
                          example: 1
                        title:
                          type: string
                          example: 'News 9'
                        content:
                          type: string
                          example: content
                        thumbnail:
                          type: string
                          example: 'http://jet-api.ethan-tech.asia/400x300.jpg'
                        policy:
                          type: string
                          example: policy
                        is_new:
                          type: boolean
                          example: true
                        is_public:
                          type: boolean
                          example: true
                        is_read:
                          type: boolean
                          example: false
                        release_time:
                          type: string
                          example: '2021-03-18 08:08:02'
                        created_at:
                          type: string
                          example: '2021-03-18 08:08:02'
                        updated_at:
                          type: string
                          example: '2021-03-18 08:08:02'
                      type: object
                  links:
                    properties:
                      first:
                        type: string
                        example: 'http:\/\/jet-api.lc:801\/api\/v1\/news?page=1'
                      last:
                        type: string
                        example: 'http:\/\/jet-api.lc:801\/api\/v1\/news?page=10'
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: 'http:\/\/jet-api.lc:801\/api\/v1\/news?page=2'
                    type: object
                  meta:
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      from:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 10
                      links:
                        type: array
                        items:
                          properties:
                            url:
                              type: string
                              example: 'http:\/\/jet-api.lc:801\/api\/v1\/news?page=1'
                            label:
                              type: string
                              example: '1'
                            active:
                              type: boolean
                              example: true
                          type: object
                    type: object
                  unread_count:
                    type: integer
                    example: 10
                type: object
    post:
      tags:
        - News
      summary: 'Create a news'
      description: 'Create a new news'
      operationId: Insert
      requestBody:
        description: 'Create a news'
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - content
                - release_time
                - thumbnail
              properties:
                title:
                  type: string
                  example: 'News title'
                content:
                  type: string
                  example: 'Suscipit earum dicta rem atque. Vero facilis inventore tempora sapiente. Labore quis id enim eligendi quo debitis. Sit rerum quam temporibus sed.'
                is_public:
                  type: boolean
                  example: false
                policy:
                  type: string
                  example: na
                thumbnail:
                  type: file
                  example: images/news/400x300.jpg
                release_time:
                  type: string
                  example: '2020-11-28 03:05:06'
              type: object
      responses:
        200:
          description: 'Insert news success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Insert news success'
                  data:
                    properties:
                      id:
                        type: integer
                        example: 2
                      title:
                        type: string
                        example: 'News title'
                      content:
                        type: string
                        example: 'Suscipit earum dicta rem atque. Vero facilis inventore tempora sapiente. Labore quis id enim eligendi quo debitis. Sit rerum quam temporibus sed.'
                      is_public:
                        type: boolean
                        example: false
                      is_new:
                        type: boolean
                        example: true
                      policy:
                        type: string
                        example: na
                      thumbnail:
                        type: file
                        example: 'localhost:801/images/news/400x300.jpg'
                      release_time:
                        type: string
                        example: '2020-11-28 03:05:06'
                      created_at:
                        type: string
                        example: '2020-11-27 03:05:06'
                      updated_at:
                        type: string
                        example: '2020-11-27 03:37:17'
                    type: object
                type: object
...
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: 'Get all notifications'
      description: 'Get all notifications'
      operationId: 'Get all'
      parameters:
        -
          name: fetch_all
          in: query
          description: 'Return all rows (fetch_all=1) or per page (default)'
          schema:
            type: string
            format: int
            example: 1
      responses:
        200:
          description: 'Get all notifications success'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Get all notifications success'
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        title:
                          type: string
                          example: 'Notifications title'
                        body:
                          type: string
                          example: 'Suscipit earum dicta rem atque. Vero facilis inventore tempora sapiente. Labore quis id enim eligendi quo debitis. Sit rerum quam temporibus sed.'
                        delivery_time:
                          type: string
                          example: '2020-11-28 03:05:06'
                        created_at:
                          type: string
                          example: '2020-11-27 03:05:06'
                        updated_at:
                          type: string
                          example: '2020-11-27 03:37:17'
                        member_notifications:
                          type: array
                          items:
                            properties:
                              notification_id:
                                type: integer
                                example: 1
                              member_id:
                                description: '0: all members | others: member_id matches the values'
                                type: integer
                                example: 1
                              status:
                                description: '0: unsent | 1: sent | 2: error'
                                type: integer
                                example: 0
                            type: object
                        is_unsent:
                          type: boolean
                          example: true
                      type: object
                  links:
                    properties:
                      first:
                        type: string
                        example: 'http:\/\/jet-api.lc:801\/api\/v1\/notifications?page=1'
                      last:
                        type: string
                        example: 'http:\/\/jet-api.lc:801\/api\/v1\/notifications?page=10'
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: 'http:\/\/jet-api.lc:801\/api\/v1\/notifications?page=2'
                    type: object
                  meta:
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      from:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 10
                      links:
                        type: array
                        items:
                          properties:
                            url:
                              type: string
                              example: 'http:\/\/jet-api.lc:801\/api\/v1\/notifications?page=1'
                            label:
                              type: string
                              example: '1'
                            active:
                              type: boolean
                              example: true
                          type: object
                    type: object
                type: object
    post:
      tags:
        - Notifications
      summary: 'Create a new notification'
      description: 'Create a new notification'
      operationId: Create
      requestBody:
        description: 'Input data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreateRequest'
      responses:
        200:
          description: 'Store notification successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResource'
  '/api/v1/notifications/{notification_id}':
    get:
      tags:
        - Notifications
      summary: 'Get a notification info by id'
      description: 'Get a notification info by id'
      operationId: Find
      parameters:
        -
          name: notification_id
          in: path
          description: 'ID of notification to return'
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        200:
          description: 'Find notification success'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResource'
    put:
      tags:
        - Notifications
      summary: 'Update a notification'
      description: 'Update a notification'
      operationId: Update
      parameters:
        -
          name: notification_id
          in: path
          description: 'ID of notification to update'
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        description: 'Input data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationUpdateRequest'
      responses:
        200:
          description: 'Update notification successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResource'
    delete:
      tags:
        - Notifications
      summary: 'Delete a notification'
      description: 'Delete a notification'
      operationId: Delete
      parameters:
        -
          name: notification_id
          in: path
          description: 'ID of notification to delete'
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        200:
          description: 'Update notification successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResource'
...
  '/api/v1/notifications/send/{member_id}':
    post:
      tags:
        - Notifications
      summary: 'Send test notification'
      description: 'Send test notification'
      operationId: SendTest
      parameters:
        -
          name: member_id
          in: path
          description: 'Member_id of the receiver'
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        description: 'Input data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSendRequest'
      responses:
        200:
          description: 'Send notification successfully'
          content:
            application/json:
              schema:
                properties:
                  status_code:
                    type: integer
                    example: '200'
                  message:
                    type: string
                    example: 'Notifications sent.'
                  data:
                    type: object
                    example: 'null'
                type: object
...
components:
  schemas:
    NotificationCreateRequest:
      required:
        - title
        - body
        - delivery_time
      properties:
        title:
          type: string
          example: 'First notification'
        body:
          type: string
          example: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.'
        delivery_time:
          type: string
          example: '2020-11-27 03:37:17'
        member_id:
          type: array
          items:
            description: '- 0: all members | others: member_id matches the values. *Required if prefecture_id not specified.'
            type: integer
            example: 1
        prefecture_id:
          type: array
          items:
            type: integer
            example: 1
      type: object
    NotificationSendRequest:
      required:
        - title
        - body
      properties:
        title:
          type: string
          example: 'First notification'
        body:
          type: string
          example: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.'
      type: object
...
    NotificationUpdateRequest:
      properties:
        title:
          type: string
          example: 'First notification'
        body:
          type: string
          example: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.'
        delivery_time:
          type: string
          example: '2020-11-27 03:37:17'
        member_id:
          type: array
          items:
            description: '- 0: all members | others: member_id matches the values. *Required if prefecture_id not specified.'
            type: integer
            example: 1
        prefecture_id:
          type: array
          items:
            type: integer
            example: 1
      type: object
...
  securitySchemes:
    bearerAuth:
      type: apiKey
      description: 'A token created from Sanctum Laravel package'
      name: Authorization
      in: header
security:
  - []